pyproject.toml
[project]
name = "jwt-authentication"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.135.1",
    "pwdlib[bcrypt]>=0.3.0",
    "pydantic>=2.12.5",
    "pyjwt>=2.11.0",
    "sqlalchemy>=2.0.48",
]
